pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cashflow"
version = "0.5.0"
description = "Cash Flow Forecasting Predictive Engine - SDD v0.05"
readme = "README.md"
license = "Unlicense"
requires-python = ">=3.8"
authors = [
    { name = "Development Team" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "statsmodels>=0.13.0",
    "scikit-learn>=1.0.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pandas-stubs>=2.0.0",
]

[project.scripts]
cashflow = "cashflow.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/cashflow"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=cashflow --cov-report=term-missing"
pythonpath = ["src"]

[tool.mypy]
python_version = "3.8"
strict = true
plugins = ["pydantic.mypy"]
mypy_path = "src"

[tool.ruff]
line-length = 100
target-version = "py38"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
